requirements.txt
Adafruit-Blinka==8.32.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-framebuf==1.6.5
adafruit-circuitpython-motor==3.4.13
adafruit-circuitpython-pca9685==3.4.14
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==2.0.5
adafruit-circuitpython-servokit==1.3.17
adafruit-circuitpython-ssd1306==2.12.16
adafruit-circuitpython-typing==1.10.2
Adafruit-PlatformDetect==3.60.0
Adafruit-PureIO==1.1.11
blinker==1.7.0
click==8.1.7
filelock==3.13.1
flask==3.0.2
flatbuffers==20181003210633
fsspec==2024.2.0
importlib-metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
lewansoul-lx16a==0.1
lewansoul-lx16a-terminal==0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
pigpio==1.78
pillow==10.2.0
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sympy==1.12
sysv-ipc==1.1.0
tflite==2.10.0
tflite-runtime==2.13.0
traitlets==5.14.1
typing-extensions==4.9.0
werkzeug==3.0.1
zipp==3.18.0
